+++ nornir_imageregistration/__init__.py
@@ -24,6 +24,15 @@ assemble_tiles
 
 '''
 
-from nornir_imageregistration.alignment_record import *
+from nornir_imageregistration.core import *
+from nornir_imageregistration.alignment_record import AlignmentRecord
+from nornir_imageregistration.volume import Volume
+from nornir_imageregistration.mosaic import Mosaic
+
+import nornir_imageregistration.files
+from nornir_imageregistration.spatial import *
+import nornir_imageregistration.transforms
+
+
 
 __all__ = ['image_stats', 'core', 'files', 'geometry', 'transforms', 'spatial']