+++ nornir_imageregistration/image_stats.py
@@ -15,12 +15,13 @@ from pylab import median, mean, std, sqrt, imread, ceil, floor, mod
 import scipy.ndimage.measurements
 import scipy.stats
 
-import core
-import im_histogram_parser
 import nornir_pools as pools
 import nornir_shared.histogram
 import nornir_shared.images as images
 import nornir_shared.prettyoutput as PrettyOutput
+
+from . import core
+from . import im_histogram_parser
 
 
 class ImageStats():
@@ -81,7 +82,7 @@ def Prune(filenames, MaxOverlap=None):
         FilenameToResult[k] = float(FilenameToResult[k][1]);
 
     if isinstance(filenames, str):
-        return FilenameToResult.items()[0];
+        return list(FilenameToResult.items())[0];
     else:
         return FilenameToResult;
 
@@ -244,7 +245,7 @@ def Histogram(filenames, Bpp=None, MinSampleCount=None, Scale=None, numBins=None
     OutputMap = {};
     minVal = None;
     maxVal = None;
-    for f in FilenameToTask.keys():
+    for f in list(FilenameToTask.keys()):
         task = FilenameToTask[f]
         taskOutput = task.wait_return();
         lines = taskOutput.splitlines();
@@ -265,7 +266,7 @@ def Histogram(filenames, Bpp=None, MinSampleCount=None, Scale=None, numBins=None
     threadTasks = [];
     TPool = pools.GetGlobalMultithreadingPool();
 
-    for f in OutputMap.keys():
+    for f in list(OutputMap.keys()):
         threadTask = TPool.add_task(f, im_histogram_parser.Parse, OutputMap[f], minVal=minVal, maxVal=maxVal, numBins=numBins);
         threadTasks.append(threadTask);
 
@@ -348,10 +349,8 @@ def __HistogramFileImageMagick__(filename, ProcPool, Bpp=None, Scale=None):
     if Scale > 1:
         Scale = 1;
 
-
     CmdTemplate = "convert %(filename)s -filter point -scale %(scale)g%% -define histogram:unique-colors=true -format %%c histogram:info:- && exit"
     Cmd = CmdTemplate % {'filename' : filename, 'scale' : Scale * 100};
-
 
     task = ProcPool.add_process(os.path.basename(filename), Cmd, shell=True);
 
@@ -390,7 +389,7 @@ if __name__ == '__main__':
         pr = pstats.Stats(ProfilePath);
         if not pr is None:
             pr.sort_stats('time');
-            print str(pr.print_stats(.05));
+            print(str(pr.print_stats(.05)));