+++ nornir_imageregistration/nornir_enum.py
@@ -7,5 +7,5 @@ Created on Jun 4, 2013
 def enum(*sequential, **named):
     '''Generates a dictionary of names to number values used as an enumeration.
        Copied from a StackOverflow post'''
-    enums = dict(zip(sequential, range(len(sequential))), **named)
+    enums = dict(list(zip(sequential, list(range(len(sequential))))), **named)
     return type('Enum', (), enums)