+++ nornir_imageregistration/tileset.py
@@ -4,22 +4,24 @@ Created on Nov 18, 2013
 @author: u0490822
 '''
 
-import numpy as np
-import core
-from scipy import stats
-# from nornir_imageregistration.files.mosaicfile import MosaicFile
-# from nornir_imageregistration.mosaic import Mosaic
-import core
-import os
-import logging
-import transforms.utils as tutils
-import nornir_pools as pools
 import copy
-from scipy.misc import imsave
+import logging
+import os
 
 import PIL
+from scipy import stats
+from scipy.misc import imsave
+
+import nornir_imageregistration.transforms.utils as tutils
+import nornir_pools as pools
+import numpy as np
+
+from . import core
+from . import core
 
 
+# from nornir_imageregistration.files.mosaicfile import MosaicFile
+# from nornir_imageregistration.mosaic import Mosaic
 class ShadeCorrectionTypes(object):
     BRIGHTFIELD = 0
     DARKFIELD = 1